scraper/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.2
playwright==1.44.0
supabase==2.4.6
python-dotenv==1.0.1
